+++ src/nksip_call_uac_resp.erl
@@ -143,7 +143,7 @@ response_status(invite_proceeding, Resp, #trans{code=Code}=UAC, Call)
     UAC1 = nksip_call_lib:timeout_timer(timer_c, UAC, Call),
     Call1 = nksip_call_uac_reply:reply({resp, Resp}, UAC1, Call),
     case Cancel of
-        to_cancel -> nksip_call_uac:cancel(UAC1, update(UAC1, Call1));
+        to_cancel -> nksip_call_uac:cancel(UAC1, undefined, update(UAC1, Call1));
         _ -> update(UAC1, Call1)
     end;