+++ doc/CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog
 * Full event support (SUBSCRIBE/NOTIFY).
 * Full PUBLISH support, using in-memory or external database.
 * Path support, as client, proxy and registrar.
+* Reason header support in request and responses. 
 * Support for headers in URIs.
 * UAS callback functions receive contextual metadata.
 * New options to customize supported extensions and to generate Require and Accept headers.